@@ -49,47 +44,46 @@
         "mocha-junit-reporter": "~1.17.0",
         "mocha-multi-reporters": "~1.1.7",
         "npm-run-all": "~4.1.3",
-        "nyc": "~11.8.0",
+        "nyc": "~12.0.2",
         "pug": "~2.0.3",
-        "query-string": "~6.1.0",
         "rimraf": "~2.6.2",
-        "rxjs": "~5.5.10",
-        "sinon": "~5.0.10",
+        "rxjs": "~6.2.2",
+        "sinon": "~4.5.0",
         "source-map-support": "~0.5.6",
-        "ts-loader": "~4.3.0",
+        "ts-loader": "~4.4.1",
         "ts-node": "~5.0.1",
-        "tslint": "~5.10.0",
+        "tslint": "~5.11.0",
         "typedoc": "~0.11.1",
-        "typescript": "~2.8.1",
-        "webpack": "~4.9.1",
+        "typescript": "~3.0.1",
+        "webpack": "~4.16.5",
         "webpack-bundle-analyzer": "~2.13.1",
-        "webpack-cli": "~2.1.4",
+        "webpack-cli": "~3.1.0",
         "webpack-glsl-loader": "~1.0.1",
-        "webpack-rxjs-externals": "~1.1.0"
+        "webpack-rxjs-externals": "~2.0.0"
     },
     "peerDependencies": {
-        "rxjs": "~5.5.10"
+        "rxjs": "~6.2.2"
     },
-    "sideEffects": [
-        "./source/polyfill.ts"
-    ],
     "scripts": {
         "build": "webpack-cli --mode development && node pugconfig.js",
-        "build-watch": "webpack-cli --mode development --watch",
-        "start": "http-server ./build -a localhost -p 4000",
+        "build-watch": "webpack-cli --mode production --watch",
+        "start": "http-server ./build -a localhost -c-1",
         "start-watch": "run-p build-watch pug-watch start",
         "build-dist": "webpack-cli --mode production --config webpack.config.dist.js",
         "build-lib": "webpack-cli --mode production --config webpack.config.lib.js",
         "pug-watch": "node pugconfig.js --watch",
         "cleanup": "rimraf ./dist ./lib ./build",
         "lint": "tslint --project tsconfig.json --force --format stylish --config tslint.json",
-        "test": "nyc mocha",
+        "test": "mocha",
+        "test-nyc": "nyc mocha",
         "test-watch": "mocha --watch",
         "coverage": "nyc report --reporter=text-lcov | coveralls",
-        "doc": "typedoc --hideGenerator --readme README.md --target ES6 --module commonjs --mode file --out dist/doc/ source/",
-        "ci": "npm run-script build-dist && npm run-script build-lib && npm run-script lint && npm run-script test",
+        "doc": "typedoc source/webgl-operate.ts",
+        "ci": "run-p build-dist build-lib lint test-nyc",
         "analyze": "cross-env ANALYZE=true webpack-cli --mode production --config webpack.config.js",
-        "prepack": "npm install && run-p build-dist build-lib"
+        "postversion": "git push && git push --tags",
+        "prepack": "npm run preversion",
+        "preversion": "npm install && npm run ci"
     },
     "keywords": [
         "webgl",
