@@ -1,19 +1,26 @@
 
 import { vec2 } from 'gl-matrix';
 
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Observable, Subscription } from 'rxjs';
 
 
 import { assert } from './auxiliaries';
 
 import { MouseEventProvider } from './mouseeventprovider';
 import { Invalidate } from './renderer';
+import { TouchEventProvider } from './toucheventprovider';
+
 
 /**
  * Callback for handling mouse events, given the latest mouse events (since last update) as well as the previous.
  */
 export interface MouseEventHandler { (latests: Array<MouseEvent>, previous: Array<MouseEvent>): void; }
+
+/**
+ * Callback for handling touch events, given the latest touch events (since last update) as well as the previous.
+ */
+export interface TouchEventHandler { (latests: Array<TouchEvent>, previous: Array<TouchEvent>): void; }
+
 
 /**
  * ... Provider and event handler are explicitly separated in order to reduce the number of observables (reuse of event
@@ -41,6 +48,11 @@ export class EventHandler {
      */
     protected _mouseEventProvider: MouseEventProvider | undefined;
 
+    /**
+     * Assigned touch event provider. This is usually created and owned by the canvas.
+     */
+    protected _touchEventProvider: TouchEventProvider | undefined;
+
     protected _latestMouseEventsByType =
         new Map<MouseEventProvider.Type, Array<MouseEvent>>();
     protected _previousMouseEventsByType =
@@ -48,16 +60,19 @@ export class EventHandler {
     protected _mouseEventHandlerByType =
         new Map<MouseEventProvider.Type, Array<MouseEventHandler>>();
 
-    /**
-     * Assigned touch event provider. This is usually created and owned by the canvas.
-     */
-    // protected _touchEventProvider: TouchEventProvider | undefined;
+    protected _latestTouchEventsByType =
+        new Map<TouchEventProvider.Type, Array<TouchEvent>>();
+    protected _previousTouchEventsByType =
+        new Map<TouchEventProvider.Type, Array<TouchEvent>>();
+    protected _touchEventHandlerByType =
+        new Map<TouchEventProvider.Type, Array<TouchEventHandler>>();
 
 
     constructor(invalidate: Invalidate | undefined, mouseEventProvider: MouseEventProvider | undefined,
-        /* keyEventProvider: KeyEventProvider | undefined, touchEventProvider: TouchEventProvider | undefined */) {
+        /* keyEventProvider: KeyEventProvider | undefined,*/ touchEventProvider: TouchEventProvider | undefined) {
         this._invalidate = invalidate;
         this._mouseEventProvider = mouseEventProvider;
+        this._touchEventProvider = touchEventProvider;
     }
 
 
@@ -65,7 +80,7 @@ export class EventHandler {
      * A callback intended to be invoked whenever the specialized event handler itself is invalid. By default only
      * optional updates (renderer decides whether or not to continue redraw) are triggered.
      */
-    protected invalidate(force: boolean = false) {
+    protected invalidate(force: boolean = false): void {
         if (this._invalidate) {
             this._invalidate(force);
         }
@@ -79,33 +94,35 @@ export class EventHandler {
      * @param type - Mouse event type the handler is to be associated with.
      * @param handler - Handler to be called on update.
      */
-    protected pushMouseEventHandler(type: MouseEventProvider.Type, handler: MouseEventHandler) {
+    protected pushMouseEventHandler(type: MouseEventProvider.Type, handler: MouseEventHandler): void {
+        if (this._mouseEventHandlerByType.has(type)) {
+            (this._mouseEventHandlerByType.get(type) as Array<MouseEventHandler>).push(handler);
+            return;
+        }
 
-        if (!this._mouseEventHandlerByType.has(type)) {
-            this._mouseEventHandlerByType.set(type, new Array<MouseEventHandler>());
+        this._mouseEventHandlerByType.set(type, new Array<MouseEventHandler>());
 
-            this._previousMouseEventsByType.set(type, new Array<MouseEvent>());
-            const latest = new Array<MouseEvent>();
-            this._latestMouseEventsByType.set(type, latest);
+        this._previousMouseEventsByType.set(type, new Array<MouseEvent>());
+        const latest = new Array<MouseEvent>();
+        this._latestMouseEventsByType.set(type, latest);
 
-            assert(this._mouseEventProvider !== undefined, `expected valid mouse event provider`);
-            const observable = (this._mouseEventProvider as MouseEventProvider).observable(type);
+        assert(this._mouseEventProvider !== undefined, `expected valid mouse event provider`);
+        const observable = (this._mouseEventProvider as MouseEventProvider).observable(type);
 
-            switch (type) {
-                case MouseEventProvider.Type.Wheel:
-                    this._subscriptions.push((observable as Observable<WheelEvent>).subscribe(
-                        (event) => { latest.push(event); this.invalidate(); }));
-                    break;
-                default:
-                    this._subscriptions.push((observable as Observable<MouseEvent>).subscribe(
-                        (event) => { latest.push(event); this.invalidate(); }));
-                    break;
-            }
+        switch (type) {
+            case MouseEventProvider.Type.Wheel:
+                this._subscriptions.push((observable as Observable<WheelEvent>).subscribe(
+                    (event) => { latest.push(event); this.invalidate(); }));
+                break;
+            default:
+                this._subscriptions.push((observable as Observable<MouseEvent>).subscribe(
+                    (event) => { latest.push(event); this.invalidate(); }));
+                break;
         }
         (this._mouseEventHandlerByType.get(type) as Array<MouseEventHandler>).push(handler);
     }
 
-    protected invokeMouseEventHandler(type: MouseEventProvider.Type) {
+    protected invokeMouseEventHandler(type: MouseEventProvider.Type): void {
         const handlers = this._mouseEventHandlerByType.get(type);
         if (handlers === undefined || handlers.length === 0) {
             return;
@@ -121,13 +138,58 @@ export class EventHandler {
         latest.length = 0;
     }
 
+    /**
+     * Utility for registering an additional touch event handler for updates on touch events of the given type. The
+     * handler is to be called on update iff at least a single touch event of the given type has occurred since last
+     * update.
+     * @param type - Touch event type the handler is to be associated with.
+     * @param handler - Handler to be called on update.
+     */
+    protected pushTouchEventHandler(type: TouchEventProvider.Type, handler: TouchEventHandler): void {
+        if (this._touchEventHandlerByType.has(type)) {
+            (this._touchEventHandlerByType.get(type) as Array<TouchEventHandler>).push(handler);
+            return;
+        }
+
+        this._touchEventHandlerByType.set(type, new Array<TouchEventHandler>());
+
+        this._previousTouchEventsByType.set(type, new Array<TouchEvent>());
+        const latest = new Array<TouchEvent>();
+        this._latestTouchEventsByType.set(type, latest);
+
+        assert(this._touchEventProvider !== undefined, `expected valid touch event provider`);
+        const observable = (this._touchEventProvider as TouchEventProvider).observable(type);
+
+        this._subscriptions.push((observable as Observable<TouchEvent>).subscribe(
+            (event) => { latest.push(event); this.invalidate(); }));
+
+        (this._touchEventHandlerByType.get(type) as Array<TouchEventHandler>).push(handler);
+    }
+
+    protected invokeTouchEventHandler(type: TouchEventProvider.Type): void {
+        const handlers = this._touchEventHandlerByType.get(type);
+        if (handlers === undefined || handlers.length === 0) {
+            return;
+        }
+        const latest = this._latestTouchEventsByType.get(type) as Array<TouchEvent>;
+        if (latest.length === 0) {
+            return;
+        }
+        const previous = this._previousTouchEventsByType.get(type) as Array<TouchEvent>;
+        handlers.forEach((handler) => handler(latest, previous));
+
+        Object.assign(previous, latest);
+        latest.length = 0;
+    }
 
     /**
      * Disposes all registered handlers of all event types.
      */
-    dispose() {
+    dispose(): void {
         this._latestMouseEventsByType.forEach((value) => value.length = 0);
         this._previousMouseEventsByType.forEach((value) => value.length = 0);
+        this._latestTouchEventsByType.forEach((value) => value.length = 0);
+        this._previousTouchEventsByType.forEach((value) => value.length = 0);
 
         for (const subscription of this._subscriptions) {
             subscription.unsubscribe();
@@ -137,7 +199,7 @@ export class EventHandler {
     /**
      * Triggers (by means of a helper function) invocation of all registered handler of all event types.
      */
-    update() {
+    update(): void {
         this.invokeMouseEventHandler(MouseEventProvider.Type.Click);
         this.invokeMouseEventHandler(MouseEventProvider.Type.Enter);
         this.invokeMouseEventHandler(MouseEventProvider.Type.Leave);
@@ -145,6 +207,11 @@ export class EventHandler {
         this.invokeMouseEventHandler(MouseEventProvider.Type.Up);
         this.invokeMouseEventHandler(MouseEventProvider.Type.Move);
         this.invokeMouseEventHandler(MouseEventProvider.Type.Wheel);
+
+        this.invokeTouchEventHandler(TouchEventProvider.Type.Start);
+        this.invokeTouchEventHandler(TouchEventProvider.Type.End);
+        this.invokeTouchEventHandler(TouchEventProvider.Type.Move);
+        this.invokeTouchEventHandler(TouchEventProvider.Type.Cancel);
     }
 
 
@@ -167,9 +234,8 @@ export class EventHandler {
 
         } else if (event instanceof TouchEvent) {
             const e = event as TouchEvent;
-            /* tslint:disable-next-line:prefer-for-of */
-            for (let i = 0; i < e.touches.length; ++i) {
-                const touch = e.touches[i];
+            for (let index = 0; index < e.touches.length; ++index) {
+                const touch = e.touches.item(index)!;
                 offsets.push(vec2.fromValues(touch.clientX, touch.clientY));
             }
         }
@@ -193,7 +259,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushClickHandler(handler: MouseEventHandler) {
+    pushClickHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Click, handler);
     }
 
@@ -202,7 +268,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseEnterHandler(handler: MouseEventHandler) {
+    pushMouseEnterHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Enter, handler);
     }
 
@@ -211,7 +277,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseLeaveHandler(handler: MouseEventHandler) {
+    pushMouseLeaveHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Leave, handler);
     }
 
@@ -220,7 +286,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseDownHandler(handler: MouseEventHandler) {
+    pushMouseDownHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Down, handler);
     }
 
@@ -229,7 +295,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseUpHandler(handler: MouseEventHandler) {
+    pushMouseUpHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Up, handler);
     }
 
@@ -238,7 +304,7 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseMoveHandler(handler: MouseEventHandler) {
+    pushMouseMoveHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Move, handler);
     }
 
@@ -247,16 +313,59 @@ export class EventHandler {
      * occurred since last update.
      * @param handler - Handler to be called on update.
      */
-    pushMouseWheelHandler(handler: MouseEventHandler) {
+    pushMouseWheelHandler(handler: MouseEventHandler): void {
         this.pushMouseEventHandler(MouseEventProvider.Type.Wheel, handler);
     }
 
+    /**
+     * Register a touch start event handler that is to be called on update iff at least a single touch start event has
+     * occurred since last update.
+     * @param handler - Handler to be called on update.
+     */
+    pushTouchStartHandler(handler: TouchEventHandler): void {
+        this.pushTouchEventHandler(TouchEventProvider.Type.Start, handler);
+    }
+
+    /**
+     * Register a touch end event handler that is to be called on update iff at least a single touch end event has
+     * occurred since last update.
+     * @param handler - Handler to be called on update.
+     */
+    pushTouchEndHandler(handler: TouchEventHandler): void {
+        this.pushTouchEventHandler(TouchEventProvider.Type.End, handler);
+    }
+
+    /**
+     * Register a touch move event handler that is to be called on update iff at least a single touch move event has
+     * occurred since last update.
+     * @param handler - Handler to be called on update.
+     */
+    pushTouchMoveHandler(handler: TouchEventHandler): void {
+        this.pushTouchEventHandler(TouchEventProvider.Type.Move, handler);
+    }
+
+    /**
+     * Register a touch cancel event handler that is to be called on update iff at least a single touch cancel event has
+     * occurred since last update.
+     * @param handler - Handler to be called on update.
+     */
+    pushTouchCancelHandler(handler: TouchEventHandler): void {
+        this.pushTouchEventHandler(TouchEventProvider.Type.Cancel, handler);
+    }
+
+
+    /**
+     * Forward pointer lock request to the mouse event provider (if one exists).
+     */
     requestPointerLock(): void {
         if (this._mouseEventProvider) {
             this._mouseEventProvider.pointerLock = true;
         }
     }
 
+    /**
+     * Forward pointer lock release request to the mouse event provider (if one exists).
+     */
     exitPointerLock(): void {
         if (this._mouseEventProvider) {
             this._mouseEventProvider.pointerLock = false;
