@@ -1,8 +1,7 @@
 
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { Observable, ReplaySubject } from 'rxjs';
 
-import { assert } from './auxiliaries';
+import { assert, bitInBitfield } from './auxiliaries';
 
 import { PointerLock } from './pointerlock';
 
@@ -42,6 +41,12 @@ export class MouseEventProvider {
     /** @see {@link pointerLock} */
     protected _pointerLockRequestPending = false;
 
+    /**
+     * This mask saves for which types of events, event.preventDefault should be called. This is useful to disallow
+     * some kinds of standard events like scrolling or clicking on links.
+     */
+    protected _preventDefaultMask: MouseEventProvider.Type;
+
     constructor(element: HTMLCanvasElement, timeframe?: number) {
         assert(element !== undefined, `expected valid canvas element on initialization, given ${element}`);
         this._element = element;
@@ -63,23 +68,60 @@ export class MouseEventProvider {
         PointerLock.request(this._element);
     }
 
-    observable(type: MouseEventProvider.Type): Observable<MouseEvent> | Observable<WheelEvent> {
-        /* tslint:disable-next-line:switch-default */
+    /**
+     * Checks whether or not to prevent the default handling of the given event. This depends on the internal
+     * `preventDefaultMask` which can be modified using `preventDefault` function @see{@link prevenDefault}.
+     * @param type - Internal event type of the incoming event.
+     * @param event - Actual event to prevent default handling on (if masked).
+     */
+    protected preventDefaultOnEvent(type: MouseEventProvider.Type, event: MouseEvent): void {
+        if (bitInBitfield(this._preventDefaultMask, type)) {
+            event.preventDefault();
+        }
+    }
+
+    /**
+     * Prevent default event handling on specific event types (using prevenDefault on the event).
+     * @param types - Event types to prevent default handling on.
+     */
+    preventDefault(...types: MouseEventProvider.Type[]): void {
+        for (const type of types) {
+            if (!bitInBitfield(this._preventDefaultMask, type)) {
+                this._preventDefaultMask |= type;
+            }
+        }
+    }
+
+    /**
+     * Allow default event handling on specific event types (not calling preventDefault on the event).
+     * @param types - Event types to allow default handling on.
+     */
+    allowDefault(...types: MouseEventProvider.Type[]): void {
+        for (const type of types) {
+            if (bitInBitfield(this._preventDefaultMask, type)) {
+                this._preventDefaultMask &= ~type;
+            }
+        }
+    }
+
+    observable(type: MouseEventProvider.Type): Observable<MouseEvent> | Observable<WheelEvent> | undefined {
         switch (type) {
             case MouseEventProvider.Type.Click:
-                return this.clickObservable;
+                return this.click$;
             case MouseEventProvider.Type.Enter:
-                return this.enterObservable;
+                return this.enter$;
             case MouseEventProvider.Type.Leave:
-                return this.leaveObservable;
+                return this.leave$;
             case MouseEventProvider.Type.Down:
-                return this.downObservable;
+                return this.down$;
             case MouseEventProvider.Type.Up:
-                return this.upObservable;
+                return this.up$;
             case MouseEventProvider.Type.Move:
-                return this.moveObservable;
+                return this.move$;
             case MouseEventProvider.Type.Wheel:
-                return this.wheelObservable;
+                return this.wheel$;
+            default:
+                return undefined;
         }
     }
 
@@ -94,70 +136,90 @@ export class MouseEventProvider {
             PointerLock.exit();
         }
     }
-
     get pointerLock(): boolean {
         return PointerLock.active(this._element);
     }
 
-    get clickObservable(): Observable<MouseEvent> {
+    get click$(): Observable<MouseEvent> {
         if (this._clickSubject === undefined) {
             this._clickSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._clickListener = (event: MouseEvent) => this._clickSubject.next(event);
-            this._element.onclick = this._clickListener;
+            this._clickListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Click, event);
+                this._clickSubject.next(event);
+            };
+            this._element.addEventListener('click', this._clickListener);
         }
         return this._clickSubject.asObservable();
     }
 
-    get enterObservable(): Observable<MouseEvent> {
+    get enter$(): Observable<MouseEvent> {
         if (this._enterSubject === undefined) {
             this._enterSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._enterListener = (event: MouseEvent) => this._enterSubject.next(event);
-            this._element.onmouseenter = this._enterListener;
+            this._enterListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Enter, event);
+                this._enterSubject.next(event);
+            };
+            this._element.addEventListener('mouseenter', this._enterListener);
         }
         return this._enterSubject.asObservable();
     }
 
-    get leaveObservable(): Observable<MouseEvent> {
+    get leave$(): Observable<MouseEvent> {
         if (this._leaveSubject === undefined) {
             this._leaveSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._leaveListener = (event: MouseEvent) => this._leaveSubject.next(event);
-            this._element.onmouseleave = this._leaveListener;
+            this._leaveListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Leave, event);
+                this._leaveSubject.next(event);
+            };
+            this._element.addEventListener('mouseleave', this._leaveListener);
         }
         return this._leaveSubject.asObservable();
     }
 
-    get downObservable(): Observable<MouseEvent> {
+    get down$(): Observable<MouseEvent> {
         if (this._downSubject === undefined) {
             this._downSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._downListener = (event: MouseEvent) => this._downSubject.next(event);
-            this._element.onmousedown = this._downListener;
+            this._downListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Down, event);
+                this._downSubject.next(event);
+            };
+            this._element.addEventListener('mousedown', this._downListener);
         }
         return this._downSubject.asObservable();
     }
 
-    get upObservable(): Observable<MouseEvent> {
+    get up$(): Observable<MouseEvent> {
         if (this._upSubject === undefined) {
             this._upSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._upListener = (event: MouseEvent) => this._upSubject.next(event);
-            this._element.onmouseup = this._upListener;
+            this._upListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Up, event);
+                this._upSubject.next(event);
+            };
+            this._element.addEventListener('mouseup', this._upListener);
         }
         return this._upSubject.asObservable();
     }
 
-    get moveObservable(): Observable<MouseEvent> {
+    get move$(): Observable<MouseEvent> {
         if (this._moveSubject === undefined) {
             this._moveSubject = new ReplaySubject<MouseEvent>(undefined, this._timeframe);
-            this._moveListener = (event: MouseEvent) => this._moveSubject.next(event);
-            this._element.onmousemove = this._moveListener;
+            this._moveListener = (event: MouseEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Move, event);
+                this._moveSubject.next(event);
+            };
+            this._element.addEventListener('mousemove', this._moveListener);
         }
         return this._moveSubject.asObservable();
     }
 
-    get wheelObservable(): Observable<WheelEvent> {
+    get wheel$(): Observable<WheelEvent> {
         if (this._wheelSubject === undefined) {
             this._wheelSubject = new ReplaySubject<WheelEvent>(undefined, this._timeframe);
-            this._wheelListener = (event: WheelEvent) => this._wheelSubject.next(event);
-            this._element.onwheel = this._wheelListener;
+            this._wheelListener = (event: WheelEvent) => {
+                this.preventDefaultOnEvent(MouseEventProvider.Type.Wheel, event);
+                this._wheelSubject.next(event);
+            };
+            this._element.addEventListener('wheel', this._wheelListener);
         }
         return this._wheelSubject.asObservable();
     }
@@ -167,6 +229,14 @@ export class MouseEventProvider {
 
 export namespace MouseEventProvider {
 
-    export enum Type { Click, Enter, Leave, Down, Up, Move, Wheel }
+    export enum Type {
+        Click = 1 << 0,
+        Wheel = 1 << 1,
+        Enter = 1 << 2,
+        Leave = 1 << 3,
+        Move = 1 << 4,
+        Down = 1 << 5,
+        Up = 1 << 6,
+    }
 
 }
