@@ -140,7 +140,7 @@ export class Buffer extends AbstractObject<WebGLBuffer> implements Bindable {
     /**
      * Returns the number of bytes this object approximately allocates on the GPU.
      */
-    get bytes() {
+    get bytes(): number {
         this.assertInitialized();
         return this.context.allocationRegister.allocated(this._identifier);
     }
@@ -149,9 +149,9 @@ export class Buffer extends AbstractObject<WebGLBuffer> implements Bindable {
      * Target to which the buffer object is bound (either GL_ARRAY_BUFFER or GL_ELEMENT_ARRAY_BUFFER).
      * Readonly access to the target (as specified on initialization) the buffer will be bound to.
      */
-    get target(): GLenum {
+    get target(): GLenum | undefined {
         this.assertInitialized();
-        return this._target as GLenum;
+        return this._target;
     }
 
 }
