@@ -12,26 +12,26 @@ export class TestNavigation {
 
 
     constructor(invalidate: Invalidate, mouseEventProvider: MouseEventProvider) {
-        this._eventHandler = new EventHandler(invalidate, mouseEventProvider);
+        this._eventHandler = new EventHandler(invalidate, mouseEventProvider, undefined);
         this._eventHandler.pushMouseEnterHandler((latests: Array<MouseEvent>, previous: Array<MouseEvent>) =>
             this.onMouseEnter(latests, previous));
         this._eventHandler.pushMouseLeaveHandler((latests: Array<MouseEvent>, previous: Array<MouseEvent>) =>
             this.onMouseLeave(latests, previous));
     }
 
-    protected onMouseEnter(latests: Array<MouseEvent>, previous: Array<MouseEvent>) {
+    protected onMouseEnter(latests: Array<MouseEvent>, previous: Array<MouseEvent>): void {
         this._altered = false;
     }
 
-    protected onMouseLeave(latests: Array<MouseEvent>, previous: Array<MouseEvent>) {
+    protected onMouseLeave(latests: Array<MouseEvent>, previous: Array<MouseEvent>): void {
         this._altered = true;
     }
 
-    update() {
+    update(): void {
         this._eventHandler.update();
     }
 
-    reset() {
+    reset(): void {
         this._altered = false;
     }
 
