@@ -2,19 +2,23 @@
 const path = require('path');
 const webpack = require('webpack');
 
+var GitRevisionPlugin = require('git-revision-webpack-plugin')
+const gitrev = new GitRevisionPlugin();
+
 const rxjsExternals = require('webpack-rxjs-externals');
 
-
 module.exports = {
-
     context: __dirname + '/source',
     cache: true,
-    entry: { 'webgl-operate': ['require.ts', 'polyfill.ts', 'webgl-operate.ts'] },
+    entry: { 'webgl-operate': ['polyfill.ts', 'webgl-operate.ts'] },
     devtool: 'source-map',
     plugins: [
         new webpack.DefinePlugin({
+            GIT_REV_VERSION: JSON.stringify(gitrev.version()),
+            GIT_REV_COMMIT: JSON.stringify(gitrev.commithash()),
+            GIT_REV_BRANCH: JSON.stringify(gitrev.branch()),
             DISABLE_ASSERTIONS: JSON.stringify(false),
-            LOG_VERBOSITY_THRESHOLD: JSON.stringify(2),
+            LOG_VERBOSITY_THRESHOLD: JSON.stringify(3),
         })
     ],
     output: {
